add button for trapper
